requirement.txt
astroid==2.2.5
autopep8==1.4.4
Django==2.2.7
djangorestframework==3.10.3
dodgy==0.1.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.5
pep8-naming==0.4.1
pkg-resources==0.0.0
prospector==1.1.7
pycodestyle==2.5.0
pydocstyle==4.0.1
pyflakes==1.6.0
pylint==2.3.1
pylint-celery==0.3
pylint-django==2.0.10
pylint-flask==0.6
pylint-plugin-utils==0.6
pytz==2019.3
PyYAML==5.1.2
requirements-detector==0.6
setoptconf==0.2.0
six==1.13.0
snowballstemmer==2.0.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
